pyproject.toml
[project]
name = "deepwiki"
version = "0.1.0"
description = ""
authors = [
    {name = "chautuankien",email = "ctkiena2@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "langgraph (>=0.4.7,<0.5.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "ipython (>=9.2.0,<10.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langchain-deepseek (>=0.1.3,<0.2.0)"
]

[tool.poetry]
packages = [{include = "deepwiki", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
